Extract translate.ts
